+++ CHANGELOG.md
@@ -1,3 +1,70 @@
+#  (2018-11-07)
+
+
+### Bug Fixes
+
+* BusRepository namespace ([300bd5e](https://gitee.com/lisgroup/vueBus/commits/300bd5e))
+* command 调用任务的方法 ([eed07d4](https://gitee.com/lisgroup/vueBus/commits/eed07d4))
+* comment ([ae5c5b4](https://gitee.com/lisgroup/vueBus/commits/ae5c5b4))
+* cookie NULL 的 bug ([ab14578](https://gitee.com/lisgroup/vueBus/commits/ab14578))
+* crontab index ([41a82b5](https://gitee.com/lisgroup/vueBus/commits/41a82b5))
+* enter 提交表单 ([5b8adf4](https://gitee.com/lisgroup/vueBus/commits/5b8adf4))
+* enter 提交表单+1s 自动消失 loading ([32043d2](https://gitee.com/lisgroup/vueBus/commits/32043d2))
+* index ([418b606](https://gitee.com/lisgroup/vueBus/commits/418b606))
+* QueryList 获取 html ([cc8f510](https://gitee.com/lisgroup/vueBus/commits/cc8f510))
+* Syntax error or access violation: 1071 ([cdd7cec](https://gitee.com/lisgroup/vueBus/commits/cdd7cec))
+* 判断是否开发模式设置不同的 URL ([f25bcff](https://gitee.com/lisgroup/vueBus/commits/f25bcff))
+* 刷新页面时加载 loading 模式 ([14252f9](https://gitee.com/lisgroup/vueBus/commits/14252f9))
+* 前端页面 ([aedb33e](https://gitee.com/lisgroup/vueBus/commits/aedb33e))
+* 取消 cli 任务 ([ff2db06](https://gitee.com/lisgroup/vueBus/commits/ff2db06))
+* 抓取列表任务完成 ([e8c900f](https://gitee.com/lisgroup/vueBus/commits/e8c900f))
+* 时区修正为 PRC ([17ff5ae](https://gitee.com/lisgroup/vueBus/commits/17ff5ae))
+* 更新任务字段修改完成 ([5f0f98a](https://gitee.com/lisgroup/vueBus/commits/5f0f98a))
+* 本地执行全部任务 ([a6fd2b8](https://gitee.com/lisgroup/vueBus/commits/a6fd2b8))
+* 本地执行全部任务逻辑 bug ([b92e599](https://gitee.com/lisgroup/vueBus/commits/b92e599))
+* 第一页任务完成 ([52c5167](https://gitee.com/lisgroup/vueBus/commits/52c5167))
+* 脚本抓取数据 ([c1312be](https://gitee.com/lisgroup/vueBus/commits/c1312be))
+* 记录数据 -- 定时任务准备 ([3567ce2](https://gitee.com/lisgroup/vueBus/commits/3567ce2))
+* 调用共用方法 ([bab9185](https://gitee.com/lisgroup/vueBus/commits/bab9185))
+
+
+### Features
+
+* A new feature task ([fd6b978](https://gitee.com/lisgroup/vueBus/commits/fd6b978))
+
+
+
+#  (2018-11-07)
+
+
+### Bug Fixes
+
+* BusRepository namespace ([300bd5e](https://gitee.com/lisgroup/vueBus/commits/300bd5e))
+* command 调用任务的方法 ([eed07d4](https://gitee.com/lisgroup/vueBus/commits/eed07d4))
+* comment ([ae5c5b4](https://gitee.com/lisgroup/vueBus/commits/ae5c5b4))
+* cookie NULL 的 bug ([ab14578](https://gitee.com/lisgroup/vueBus/commits/ab14578))
+* crontab index ([41a82b5](https://gitee.com/lisgroup/vueBus/commits/41a82b5))
+* enter 提交表单 ([5b8adf4](https://gitee.com/lisgroup/vueBus/commits/5b8adf4))
+* enter 提交表单+1s 自动消失 loading ([32043d2](https://gitee.com/lisgroup/vueBus/commits/32043d2))
+* index ([418b606](https://gitee.com/lisgroup/vueBus/commits/418b606))
+* QueryList 获取 html ([cc8f510](https://gitee.com/lisgroup/vueBus/commits/cc8f510))
+* Syntax error or access violation: 1071 ([cdd7cec](https://gitee.com/lisgroup/vueBus/commits/cdd7cec))
+* 判断是否开发模式设置不同的 URL ([f25bcff](https://gitee.com/lisgroup/vueBus/commits/f25bcff))
+* 刷新页面时加载 loading 模式 ([14252f9](https://gitee.com/lisgroup/vueBus/commits/14252f9))
+* 前端页面 ([aedb33e](https://gitee.com/lisgroup/vueBus/commits/aedb33e))
+* 取消 cli 任务 ([ff2db06](https://gitee.com/lisgroup/vueBus/commits/ff2db06))
+* 抓取列表任务完成 ([e8c900f](https://gitee.com/lisgroup/vueBus/commits/e8c900f))
+* 时区修正为 PRC ([17ff5ae](https://gitee.com/lisgroup/vueBus/commits/17ff5ae))
+* 更新任务字段修改完成 ([5f0f98a](https://gitee.com/lisgroup/vueBus/commits/5f0f98a))
+* 本地执行全部任务 ([a6fd2b8](https://gitee.com/lisgroup/vueBus/commits/a6fd2b8))
+* 本地执行全部任务逻辑 bug ([b92e599](https://gitee.com/lisgroup/vueBus/commits/b92e599))
+* 第一页任务完成 ([52c5167](https://gitee.com/lisgroup/vueBus/commits/52c5167))
+* 脚本抓取数据 ([c1312be](https://gitee.com/lisgroup/vueBus/commits/c1312be))
+* 记录数据 -- 定时任务准备 ([3567ce2](https://gitee.com/lisgroup/vueBus/commits/3567ce2))
+* 调用共用方法 ([bab9185](https://gitee.com/lisgroup/vueBus/commits/bab9185))
+
+
+
 <a name=""></a>
 #  (2018-10-21)